[2] - Changed the base of the current branch with command 'rebase' and resolved conflicts. Also added at FakeUserDataSource.php the option of getting empty list by random.
