+++ app/Infrastructure/Providers/AppServiceProvider.php
@@ -3,9 +3,7 @@
 namespace App\Infrastructure\Providers;
 
 use App\Application\UserDataSource\FakeUserDataSource;
-use App\Application\UserDataSource\FakeDataSource;
 use App\Application\UserDataSource\UserDataSource;
-use App\DataSource\Database\EloquentUserDataSource;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -29,9 +27,6 @@ class AppServiceProvider extends ServiceProvider
     {
         $this->app->bind(UserDataSource::class, function () {
             return new FakeUserDataSource();
-        });
-        $this->app->bind(UserDataSource::class, function () {
-            return new FakeDataSource();
         });
     }
 }
